backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.31
pydantic==2.8.2
PyJWT==2.9.0
python-multipart==0.0.9

